Skip missing optional cut modalities

+++ src/preprocess_pipeline/behavior/preprocess_cut.py
@@ -109,25 +109,29 @@ def run_preprocess_cut(userID, expID,pre_time,post_time):
         with open(os.path.join(exp_dir_processed_cut,iS2P_file[0:7]+'_Spikes_cut.pickle'), 'wb') as f: pickle.dump(s2p_Spikes_cut, f)
 
     ### Cut ephys data ###
-    ephys_combined = np.load(os.path.join(exp_dir_processed_recordings,'ephys.npy'))
-    # loop through all trials collecting the ephys traces
-    ephys_cut = {}
-    for iTrial in range(all_trials.shape[0]):
-        trial_onset_time = all_trials.loc[iTrial,'time']
-        trial_end_time = trial_onset_time + all_trials.loc[iTrial,'duration']
-        # collect samples from ephys
-        first_sample = np.argmax(ephys_combined[0,:] >= (trial_onset_time-pre_time))
-        last_sample = np.argmax(ephys_combined[0,:] >= (trial_end_time+pre_time))
-        if iTrial == 0:
-            ephys_cut['0'] = ephys_combined[np.newaxis,1,first_sample:last_sample]
-            ephys_cut['1'] = ephys_combined[np.newaxis,2,first_sample:last_sample]
-        else:
-            ephys_cut['0'] = sparse_cat_np(ephys_cut['0'],ephys_combined[1,first_sample:last_sample])
-            ephys_cut['1'] = sparse_cat_np(ephys_cut['1'],ephys_combined[2,first_sample:last_sample])
-    
-    ephys_cut['t'] = np.linspace(0,ephys_cut['0'].shape[1]/1000,ephys_cut['0'].shape[1])-pre_time
-    # save in pickle
-    with open(os.path.join(exp_dir_processed_cut,'ephys_cut.pickle'), 'wb') as f: pickle.dump(ephys_cut, f)
+    ephys_path = os.path.join(exp_dir_processed_recordings,'ephys.npy')
+    if os.path.exists(ephys_path):
+        ephys_combined = np.load(ephys_path)
+        # loop through all trials collecting the ephys traces
+        ephys_cut = {}
+        for iTrial in range(all_trials.shape[0]):
+            trial_onset_time = all_trials.loc[iTrial,'time']
+            trial_end_time = trial_onset_time + all_trials.loc[iTrial,'duration']
+            # collect samples from ephys
+            first_sample = np.argmax(ephys_combined[0,:] >= (trial_onset_time-pre_time))
+            last_sample = np.argmax(ephys_combined[0,:] >= (trial_end_time+pre_time))
+            if iTrial == 0:
+                ephys_cut['0'] = ephys_combined[np.newaxis,1,first_sample:last_sample]
+                ephys_cut['1'] = ephys_combined[np.newaxis,2,first_sample:last_sample]
+            else:
+                ephys_cut['0'] = sparse_cat_np(ephys_cut['0'],ephys_combined[1,first_sample:last_sample])
+                ephys_cut['1'] = sparse_cat_np(ephys_cut['1'],ephys_combined[2,first_sample:last_sample])
+        
+        ephys_cut['t'] = np.linspace(0,ephys_cut['0'].shape[1]/1000,ephys_cut['0'].shape[1])-pre_time
+        # save in pickle
+        with open(os.path.join(exp_dir_processed_cut,'ephys_cut.pickle'), 'wb') as f: pickle.dump(ephys_cut, f)
+    else:
+        print(f'Ephys data not found, skipping ephys cut: {ephys_path}')
 
     ### Cut eye data ###
     # check DLC data exists
@@ -206,27 +210,31 @@ def run_preprocess_cut(userID, expID,pre_time,post_time):
     # wheel['position']
     # wheel['speed']
     # wheel['t']
-    wheel = pickle.load(open(os.path.join(exp_dir_processed_recordings,'wheel.pickle'), "rb"))
-    # loop through all trials collecting the wheel traces
-    wheel_cut = {}
-    for iTrial in range(all_trials.shape[0]):
-        trial_onset_time = all_trials.loc[iTrial,'time']
-        trial_end_time = trial_onset_time + all_trials.loc[iTrial,'duration']
-        # collect samples from wheel
-        first_sample = np.argmax(wheel['t'] >= trial_onset_time - pre_time)
-        last_sample = np.argmax(wheel['t'] >= trial_end_time + post_time)
-        if iTrial == 0:
-            wheel_cut['position'] = wheel['position'][np.newaxis,first_sample:last_sample]
-            wheel_cut['speed'] = wheel['speed'][np.newaxis,first_sample:last_sample]
-        else:
-            wheel_cut['position'] = sparse_cat_np(wheel_cut['position'],wheel['position'][np.newaxis,first_sample:last_sample])
-            wheel_cut['speed'] = sparse_cat_np(wheel_cut['speed'],wheel['speed'][np.newaxis,first_sample:last_sample])
-    
-    # make time vector
-    wheel_sample_rate = 1/np.round(wheel['t'][1]-wheel['t'][0],4)
-    wheel_cut['t'] = np.linspace(0,wheel_cut['position'].shape[1]/wheel_sample_rate,wheel_cut['position'].shape[1]) - pre_time
-    # save in pickle
-    with open(os.path.join(exp_dir_processed_cut,'wheel.pickle'), 'wb') as f: pickle.dump(wheel_cut, f)   
+    wheel_path = os.path.join(exp_dir_processed_recordings,'wheel.pickle')
+    if os.path.exists(wheel_path):
+        wheel = pickle.load(open(wheel_path, "rb"))
+        # loop through all trials collecting the wheel traces
+        wheel_cut = {}
+        for iTrial in range(all_trials.shape[0]):
+            trial_onset_time = all_trials.loc[iTrial,'time']
+            trial_end_time = trial_onset_time + all_trials.loc[iTrial,'duration']
+            # collect samples from wheel
+            first_sample = np.argmax(wheel['t'] >= trial_onset_time - pre_time)
+            last_sample = np.argmax(wheel['t'] >= trial_end_time + post_time)
+            if iTrial == 0:
+                wheel_cut['position'] = wheel['position'][np.newaxis,first_sample:last_sample]
+                wheel_cut['speed'] = wheel['speed'][np.newaxis,first_sample:last_sample]
+            else:
+                wheel_cut['position'] = sparse_cat_np(wheel_cut['position'],wheel['position'][np.newaxis,first_sample:last_sample])
+                wheel_cut['speed'] = sparse_cat_np(wheel_cut['speed'],wheel['speed'][np.newaxis,first_sample:last_sample])
+        
+        # make time vector
+        wheel_sample_rate = 1/np.round(wheel['t'][1]-wheel['t'][0],4)
+        wheel_cut['t'] = np.linspace(0,wheel_cut['position'].shape[1]/wheel_sample_rate,wheel_cut['position'].shape[1]) - pre_time
+        # save in pickle
+        with open(os.path.join(exp_dir_processed_cut,'wheel.pickle'), 'wb') as f: pickle.dump(wheel_cut, f)
+    else:
+        print(f'Wheel data not found, skipping wheel cut: {wheel_path}')
 
     print('done')
